+++ cppcryptfs/cppcryptfs.cpp
@@ -77,13 +77,10 @@ CcppcryptfsApp::CcppcryptfsApp()
 
 	m_mountedLetters = 0;
 
-	// get an OpenSSL EVP context to force detection of AES-NI instructions
+	// get a shared ptr to an OpenSSL EVP context to force detection of AES-NI instructions
 	// so we can use AES-NI even if EVP is never used
 
-	void *context = get_crypt_context(BLOCK_IV_LEN, AES_MODE_GCM);
-
-	if (context)
-		free_crypt_context(context);
+	auto context = get_crypt_context(BLOCK_IV_LEN, AES_MODE_GCM);	
 }