+++ cppcryptfs/ui/SettingsPropertyPage.cpp
@@ -77,12 +77,15 @@ BEGIN_MESSAGE_MAP(CSettingsPropertyPage, CPropertyPage)
 	ON_BN_CLICKED(IDC_OPEN_ON_MOUNTING, &CSettingsPropertyPage::OnClickedOpenOnMounting)
 	ON_BN_CLICKED(IDC_ENCRYPT_KEYS_IN_MEMORY, &CSettingsPropertyPage::OnClickedEncryptKeysInMemory)
 	ON_BN_CLICKED(IDC_CACHE_KEYS_IN_MEMORY, &CSettingsPropertyPage::OnClickedCacheKeysInMemory)
+	ON_BN_CLICKED(IDC_FAST_MOUNTING, &CSettingsPropertyPage::OnBnClickedFastMounting)
 END_MESSAGE_MAP()
 
 
 // CSettingsPropertyPage message handlers
 
-static int buffer_sizes[] = { 4, 8, 16, 32, 64, 128, 256, 512, 1024 };
+typedef int buffer_size_t;
+
+static buffer_size_t buffer_sizes[] = { 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, MAX_IO_BUFFER_KB };
 
 static int ttls[] = { 0, 1, 2, 5, 10, 15, 30, 45, 60, 90, 120, 300, 600, 900, 1800, 3600};
 
@@ -100,6 +103,15 @@ BOOL CSettingsPropertyPage::OnInitDialog()
 	int nThreads = theApp.GetProfileInt(L"Settings", L"Threads", PER_FILESYSTEM_THREADS_DEFAULT);
 
 	int bufferblocks = theApp.GetProfileInt(L"Settings", L"BufferBlocks", BUFFERBLOCKS_DEFAULT);
+
+	buffer_size_t kb = bufferblocks * 4;
+	
+	auto p = bsearch(&kb, buffer_sizes, _countof(buffer_sizes), sizeof(buffer_size_t), [](const void* pkey, const void* pval) -> int {
+		return *reinterpret_cast<const buffer_size_t*>(pkey) - *reinterpret_cast<const buffer_size_t*>(pval);
+	});
+
+	if (!p)
+		bufferblocks = BUFFERBLOCKS_DEFAULT;
 
 	int cachettl = theApp.GetProfileInt(L"Settings", L"CacheTTL", CACHETTL_DEFAULT);
 
@@ -124,15 +136,19 @@ BOOL CSettingsPropertyPage::OnInitDialog()
 	bool bCacheKeysInMemory = theApp.GetProfileIntW(L"Settings", L"CacheKeysInMemory",
 											CACHE_KEYS_IN_MEMORY_DEFAULT) != 0;
 
+	bool bFastMounting = theApp.GetProfileIntW(L"Settings", L"FastMounting",
+		FAST_MOUNTING_DEFAULT) != 0;
+
+
 	return SetControls(nThreads, bufferblocks, cachettl, bCaseInsensitive, bMountManager, 
 								bEnableSavingPasswords, bNeverSaveHistory, bDeleteSpurriousFiles, bOpenOnMounting,
-								bEncryptKeysInMemory, bCacheKeysInMemory);
+								bEncryptKeysInMemory, bCacheKeysInMemory, bFastMounting);
 }
 
 BOOL CSettingsPropertyPage::SetControls(int nThreads, int bufferblocks, int cachettl, 
 						bool bCaseInsensitive, bool bMountManager, bool bEnableSavingPasswords, bool bNeverSaveHistory,
 						bool bDeleteSpurriousFiles, bool bOpenOnMounting, bool bEncryptKeysInMemory,
-						bool bCacheKeysInMemory)
+						bool bCacheKeysInMemory, bool bFastMounting)
 {
 
 	m_bCaseInsensitive =  bCaseInsensitive;
@@ -143,6 +159,7 @@ BOOL CSettingsPropertyPage::SetControls(int nThreads, int bufferblocks, int cach
 	m_bOpenOnMounting = bOpenOnMounting;
 	m_bEncryptKeysInMemory = bEncryptKeysInMemory;
 	m_bCacheKeysInMemory = bCacheKeysInMemory;
+	m_bFastMounting = bFastMounting;
 
 	int i;
 
@@ -155,7 +172,10 @@ BOOL CSettingsPropertyPage::SetControls(int nThreads, int bufferblocks, int cach
 
 	WCHAR buf[80];
 
-	for (i = 0; i < 15; i++) {
+#define DOKAN_MAX_THREAD 63 // this is defined in a header file that Dokany doesn't distribute
+
+	for (i = 0; i <= DOKAN_MAX_THREAD; i++) {
+		
 		if (i == 0) {
 			CString def = L"Dokany default (";
 			WCHAR buf[32];
@@ -230,6 +250,8 @@ BOOL CSettingsPropertyPage::SetControls(int nThreads, int bufferblocks, int cach
 	CheckDlgButton(IDC_ENCRYPT_KEYS_IN_MEMORY, m_bEncryptKeysInMemory ? 1 : 0);
 
 	CheckDlgButton(IDC_CACHE_KEYS_IN_MEMORY, m_bCacheKeysInMemory ? 1 : 0);
+
+	CheckDlgButton(IDC_FAST_MOUNTING, m_bFastMounting ? 1 : 0);
 
 	return TRUE;  // return TRUE unless you set the focus to a control
 				  // EXCEPTION: OCX Property Pages should return FALSE
@@ -310,6 +332,7 @@ void CSettingsPropertyPage::SaveSettings()
 	m_bOpenOnMounting = !m_bOpenOnMounting; // ditto
 	m_bEncryptKeysInMemory = !m_bEncryptKeysInMemory; // ditto
 	m_bCacheKeysInMemory = !m_bCacheKeysInMemory; // ditto
+	m_bFastMounting = !m_bFastMounting; // ditto
 
 	OnBnClickedCaseinsensitive();
 	OnClickedMountmanager();
@@ -319,6 +342,7 @@ void CSettingsPropertyPage::SaveSettings()
 	OnClickedOpenOnMounting();
 	OnClickedEncryptKeysInMemory();
 	OnClickedCacheKeysInMemory();
+	OnBnClickedFastMounting();
 }
 
 void CSettingsPropertyPage::OnBnClickedDefaults()
@@ -328,7 +352,7 @@ void CSettingsPropertyPage::OnBnClickedDefaults()
 	SetControls(PER_FILESYSTEM_THREADS_DEFAULT, BUFFERBLOCKS_DEFAULT, CACHETTL_DEFAULT, 
 		CASEINSENSITIVE_DEFAULT, MOUNTMANAGER_DEFAULT, ENABLE_SAVING_PASSWORDS_DEFAULT,
 		NEVER_SAVE_HISTORY_DEFAULT, DELETE_SPURRIOUS_FILES_DEFAULT, OPEN_ON_MOUNTING_DEFAULT,
-		ENCRYPT_KEYS_IN_MEMORY_DEFAULT, CACHE_KEYS_IN_MEMORY_DEFAULT);
+		ENCRYPT_KEYS_IN_MEMORY_DEFAULT, CACHE_KEYS_IN_MEMORY_DEFAULT, FAST_MOUNTING_DEFAULT);
 
 	SaveSettings();
 }
@@ -341,7 +365,7 @@ void CSettingsPropertyPage::OnBnClickedRecommended()
 	SetControls(PER_FILESYSTEM_THREADS_RECOMMENDED, BUFFERBLOCKS_RECOMMENDED, CACHETTL_RECOMMENDED, 
 		CASEINSENSITIVE_RECOMMENDED, MOUNTMANAGER_RECOMMENDED, ENABLE_SAVING_PASSWORDS_RECOMMENDED,
 		NEVER_SAVE_HISTORY_RECOMMENDED, DELETE_SUPRRIOUS_FILES_RECOMMENDED, OPEN_ON_MOUNTING_RECOMMENDED,
-		ENCRYPT_KEYS_IN_MEMORY_RECOMMENDED, CACHE_KEYS_IN_MEMORY_RECOMMENDED);
+		ENCRYPT_KEYS_IN_MEMORY_RECOMMENDED, CACHE_KEYS_IN_MEMORY_RECOMMENDED, FAST_MOUNTING_RECOMMENDED);
 
 	SaveSettings();
 }
@@ -488,4 +512,16 @@ void CSettingsPropertyPage::OnClickedCacheKeysInMemory()
 	CheckDlgButton(IDC_CACHE_KEYS_IN_MEMORY, m_bCacheKeysInMemory ? 1 : 0);
 
 	theApp.WriteProfileInt(L"Settings", L"CacheKeysInMemory", m_bCacheKeysInMemory ? 1 : 0);
+}
+
+
+void CSettingsPropertyPage::OnBnClickedFastMounting()
+{
+	// TODO: Add your control notification handler code here
+
+	m_bFastMounting = !m_bFastMounting;
+
+	CheckDlgButton(IDC_FAST_MOUNTING, m_bFastMounting ? 1 : 0);
+
+	theApp.WriteProfileInt(L"Settings", L"FastMounting", m_bFastMounting ? 1 : 0);
 }